+++ package.json
@@ -61,7 +61,7 @@
     "@oclif/command": "^1.0",
     "@oclif/config": "^1.12",
     "@oclif/plugin-help": "^2.0",
-    "@stoplight/json": "^2.3.1",
+    "@stoplight/json": "^2.3.2",
     "@stoplight/json-ref-resolver": "^2.0.2",
     "@stoplight/path": "^1.0.2",
     "@stoplight/types": "^9.1.2",
